+++ frontend/src/features/layersSelector/overlays/OverlayContent.tsx
@@ -1,6 +1,9 @@
+import { useAppDispatch } from '@hooks/useAppDispatch'
+import { MonitorEnvLayers } from 'domain/entities/layers/constants'
 import styled from 'styled-components'
 
 import { getName, getType } from './utils'
+import { openAMPMetadataPanel, openRegulatoryMetadataPanel } from '../metadataPanel/slice'
 import { LayerLegend } from '../utils/LayerLegend.style'
 
 import type { OverlayItem } from 'domain/types/map'
@@ -10,30 +13,35 @@ type OverlayContentProps = {
 }
 
 export function OverlayContent({ items }: OverlayContentProps) {
+  const dispatch = useAppDispatch()
+  const handleClick = (type, id) => () => {
+    if (type === MonitorEnvLayers.AMP) {
+      dispatch(openAMPMetadataPanel(id))
+    }
+    if (type === MonitorEnvLayers.REGULATORY_ENV) {
+      dispatch(openRegulatoryMetadataPanel(id))
+    }
+  }
+
   return (
     <Layerlist>
-      {items?.slice(0, 3).map(item => {
+      {items?.map(item => {
         const name = getName(item.properties, item.layerType)
         const type = getType(item.properties, item.layerType)
 
         return (
-          <LayerItem key={item.properties.id}>
+          <LayerItem key={item.properties.id} onClick={handleClick(item.layerType, item.properties.id)}>
             <LayerLegend layerType={item.layerType} name={name} type={type} />
             <Name>{name}</Name>
             <Type> / {type}</Type>
           </LayerItem>
         )
       })}
-      {items?.length === 4 && <More>1 autre zone</More>}
-      {items && items.length > 4 && <More>{items.length - 2} autres zones</More>}
     </Layerlist>
   )
 }
 
 const Layerlist = styled.ul`
-  min-width: 200px;
-  min-height: 100px;
-  border: 1px solid red;
   list-style: none;
   padding: 0;
   margin: 0;
@@ -52,10 +60,4 @@ const Name = styled.span`
 const Type = styled.span`
   color: ${p => p.theme.color.gunMetal};
   font: normal normal normal 13px/18px Marianne;
-`
-const More = styled.li`
-  background-color: white;
-  color: ${p => p.theme.color.slateGray};
-  font: italic normal bold 13px/18px Marianne;
-  padding: 5px;
 `
