+++ frontend/src/features/layersSelector/overlays/PinnedOverlay.tsx
@@ -1,0 +1,37 @@
+import { useAppDispatch } from '@hooks/useAppDispatch'
+import { IconButton, Icon, Size, Accent } from '@mtes-mct/monitor-ui'
+import styled from 'styled-components'
+
+import { OverlayContent } from './OverlayContent'
+import { closeLayerOverlay } from '../metadataPanel/slice'
+
+import type { OverlayItem } from 'domain/types/map'
+
+export function PinnedOverlay({ items }: { items: OverlayItem[] }) {
+  const dispatch = useAppDispatch()
+  const handleClick = () => {
+    dispatch(closeLayerOverlay())
+  }
+
+  return (
+    <Card>
+      <Header>
+        {items.length} zones superposées sur ce point{' '}
+        <IconButton accent={Accent.TERTIARY} Icon={Icon.Close} onClick={handleClick} size={Size.SMALL} />
+      </Header>
+      <OverlayContent items={items} />
+    </Card>
+  )
+}
+
+const Card = styled.div`
+  width: 440px;
+`
+const Header = styled.div`
+  display: flex;
+  justify-content: space-between;
+  background-color: ${p => p.theme.color.lightGray};
+  color: ${p => p.theme.color.gunMetal};
+  font: normal normal medium 13px/18px Marianne;
+  padding: 5px;
+`
